requirements.txt
python-docx
transformers
faiss-cpu
haystack-ai
torch
numpy